QS60Style: QMenu behaves badly with a lot of menu items
Due to recent change in QS60Style where combobox menu was changed to
popup menu, instead of dropdown menu, long QMenus now work sloppily.
This is due to that pixel metric PE_MenuScrollerHeight was set to zero
to avoid "non-drawn" areas in combobox popups. Of course, QMenu
scrolling really needs the scrolling area if it is to make any
scrolling.

As a solution, scroller height is now zero for combobox popup and
in QMenu it is the height of itemview item (for "standard" content),
but no any scroller indication is drawn. The menu scrolls automatically
when bottom/top part of menu list is highlighted.

Task-number: QTBUG-10073
Reviewed-by: Janne Koskinen