Fix tab stop in QPainter::drawText()
Even though we'd set the local tabstops variable to eight times the
width of 'x', we'd never use this value anywhere, causing the text
layout to use a default value of 80. This looked okay with some font
sizes, but since it did not depend on the size of font, the tab in
a large font would be too small and potentially the same size or
smaller than a single character/space. We enable setting the tabstops
on the layout whenever it hasn't been explicitly set on the QTextOption.

Task-number: QTBUG-1638
Reviewed-by: Trond