Unable to load plugin DLLs compiled with the '/clr' option
Before loading the plugins, to make sure we are loading a valid one,
we call qt_plugin_query_verification_data() and verify the version
information. The DLLS built with '/clr' compiler flag, needs full
initialization before a function can be called.
With a prior patch, as an optimaization we had changed LoadLibrary()
to LoadLibraryEx(), this will not initialize or load the dependencies,
so it crashed when calling that function. Now this exception is
caught and LoadLibrary(equivalent) is called.

Task-number: QTBUG-10012
Reviewed-by: Jan-Arve