Fix LeftToRight + setSpacing + setWrapping when drag-and-dropping
Quite a number of assumptions in qlistview assume either wrapping or spacing,
but not both.  This fixes the Flow left to right case.

The TopToBottom case is not affected.  TopToBottom + wrapping still works
TopToBottom + spacing still works, TopToBottom + spacing + wrapping is still
as broken as ever.

Drag move prior to this patch was correct, but also need the drop
to use the spacing when determining what index to apply to, not
just the drag move.

Reviewed-by: gabi
Merge-Request: 586
Task-number: QTBUG-8938