Removed fullscreen responsiveness of softkeys
The CEikCba class casting is not binary compatible between at least
S60 3.2 and 5.0 platforms, even though it is source compatible.
The removed code was causing stack corruption on emulator
when binaries were built on S60 3.2 SDK and run on S60 5.0 SDK.

Another solution for this feature needs to be found.

Task-number: QTBUG-10199
Reviewed-by: Shane Kearns
Reviewed-by: Janne Anttila