Ensured that WA_InputMethodEnabled was set before FocusInEvent was sent.
When focusInEvent is called we need to ensure that WA_InputMethodEnabled
is set, since FocusInEvent might lead to that
QInputContext::setFocusWidget is called (with the QGraphicsView as the
widget parameter).
However, if the widget still does not have WA_InputMethodEnabled set
yet it will assert in QInputContext::setFocusWidget (it requires that
the widget has the flag set).

This was a problem for the morbit bridge, where Dui requires the
attribute to be set (while Orbit actually does not require)

Task-number: QTBUG-10276
Reviewed-by: jasplin
Reviewed-by: Alexis Menard