Fixed incorrect runtime platform version check in Phonon MMF backend
In S60 3.2, CMdaAudioPlayerUtility::SetVolume(TInt) was changed from
having a void return type to returning TInt.  The code in the Phonon
MMF backend which calls this function uses a runtime platform version
check to ensure that only on S60 3.2 and above is the return value
from SetVolume treated as valid.

This check was previously testing for the wrong platform version (5.0
rather than 3.2).

Reviewed-by: Shane Kearns