Event dispatcher slow down using delays rather than thread priority
The Symbian event dispatcher has a mechanism to slow down the Qt app
to prevent viewsrv crashes and keep the device responsive. This was
implemented using a thread priority drop. But that has some bad side
effects, such as app and system performance instability.

This new implementation of a slow down mechanism uses a separate low
priority thread to test when the system is getting too busy. Adaptive
millisecond waits are used to slow the app down just enough to let the
low prioirity thread to run.

In practice this avoids the performance instability of the previous
method, and results in much better system stability where the system
stays more responsive with fewer viewsrv panics when heavy Qt apps are
running.

The slow down code kicks in after 2 seconds of busy time. The delays
grow 1ms at a time to a maximum of 1/4 of average event run time.

This updated version of the fix used an RSemaphore rather than
RFastlock, which should not have been used for cross-thread signalling.

Task-number: QTBUG-9489
Reviewed-by: Shane Kearns