bye bye QMakeProjectEnv
qmake variables would have been exported to the command run by
$$system() and - optionally - to the command run by system().
however, this was a unix-only feature and made the kernel barf
at the huge environment on older linuxes.

as we don't like platform-specific hacks which are unreliable,
in particular when a workaround exists (the commands execute
shell code after all, so one can inject arbitrary env variables),
just blow it away - it was undocumented, after all.

Reviewed-by: joerg