QS60Style: QComboBoxPrivateScroller draws holes into menu canvas
QComboBoxPrivateScroller calls eraserects for menu scroller areas,
expecting the style to fill the gaps. Unfortunately, the QS60Style
does not draw menu scrollers at all
(but they are functional when interacted with).
Therefore, these areas are now drawn as holes.

As a fix, the eraseRect call from provate class
QComboBoxPrivateScroller is removed when "s60 flag" is defined.

Task-number: QTBUG-10371
Reviewed-by: Janne Koskinen