Remove QDeclarativeLoader::resizeMode
The explicit resizeMode is superfluous, all usecases could be achieved
with the corrected default behaviour of the former ResizeLoaderToItem
mode. The NoResize usecase is covered by wrapping the loader in an extra
item.

That means: A loader automatically gets its size from the item loaded
unless the loader has an explicit size itself.

Go-ahead-by: Michael Brasser