QS60Style: QTabWidget icon size property doesn't work
QS60Style did check if the tab icon size was larger than pixel metric
for default tab icon size. If it was, then it reset the icon size to
match pixel metric. It was impossible to set large icon to tab widget.

As a fix, this check was removed from style.

Task-number: QTBUG-3102
Reviewed-by: Jason Barron