Cocoa: QShortcut(Qt::Key_Enter) does not show in the menu bar
The reason is that cocoa expects the unicode control character for
"enter", rather than the glyph showing an arrow. So rather than
using the glyph-table from QKeySequence, we should use a table
containing a mapping from Qt::Key to Cocoa control key. Luckily,
such a table exists allready in qt_cocoa_helpers, but reversed (
cocoa key to Qt::Key). So this patch creates a function that
binds all this together.

Reviewed-by: cduclos