Speedup activation of signals
The virtual QObject::qt_metacall will recurse to the whole object hierarchy
to find from which class a function should be called.
But it is possible to know, at connection time, from which exact
QMetaObject a function belongs, and the relative offset into it.
So we make the slot calls from the qt_static_metacall function.
So activation of signals is faster.

 - We must not call a slot from a class that has been destroyed.
   To avoid this, there is a check on the methodOffset. If it is
   smaller, that means we might be called (indirectly) from
   the destructor. We fallback to the virtual call to qt_metacall
   that does the right thing.

 - The signature of the static method is void (*) (QObject*,MetaCall,int,void**)
   It returns void, so the compiler is allowed to do tail recusive
   optimization.  Having the QObject* as first parameter make it ready
   on the stack for the call to the member function.

 - The new static method has to be a member function in order to
   be able to access the private slots.

Reviewed-by: brad