Improve QUrl handling of local file paths
Add QUrl::isLocalFile for a faster and more consistent checking of
whether the URL is local or not.

Improve the documentation to indicate that QUrl always treats SMB-like
file paths as local, even if the system cannot open them
(non-Windows).

Add a test to ensure that "FILE:/a.txt" is considered local too (RFC
3986 requires schemes to be interpreted in case-insensitive fashion).

Remove broken code that supported empty schemes as local file paths.

Reviewed-by: Markus Goetz