Recalculate script item widths when forcing justification
When going through the special WebKit code path for text drawing and
forcing justification for text, we would set the justification in the
text engine, but never actually update the widths of the script items.
Since the width of the script item is used to position the text, the
text would be drawn with the correct justification, but the position
would be set based on the non-justified text width. The result was
overlapping text whenever the script of the text changed. The fix goes
through the justified glyphs and sets the width and position based on the
new effective advance.

Task-number: QTBUG-10421
Reviewed-by: Simon Hausmann