Fix thread synchronization issues in Symbian QFileSystemWatcher
In Symbian the QSymbianFileSystemWatcherEngine thread now stays running
as long as the instance is alive to avoid repeatedly stopping and
restarting the thread as watched paths are removed and added.

Also fixed issue of misreporting adding failure in cases where both
adds and removes were done in quick succession.

Task-number: QTBUG-10091
Reviewed-by: Shane Kearns