Check for existance of sis file parameter to runonphone
Previously, it would skip over a non existant sis file - which could
cause invalid test results if a previous version of the sis was already
installed.
Now, it checks if the file exists and exits with an error message if
asked to install a sis file that doesn't exist.

Task-number: QTBUG-9290
Reviewed-by: Miikka Heikkinen