unify QMAKE_QMAKE path separator fixing
the value of the variable in Option is only ever accessed via the
project variable, so there is no point in early fixing.

as it happens, this fixes mingw+sh generating makefiles with the wrong
separator, as the fixing is delayed to a point where QMAKE_DIR_SEP was
read back into Option.

Reviewed-by: joerg