QLineEdit / QDateEdit - white font on white background
Currently QCoeFepInputContext uses default QLineEdit palette when
setting up text format before passing the formatted text to native
side (which uses it, for example, to show T9 suggested words).

The above way is incorrect due to two reasons:
- custom widget might not be anything like QLineEdit and might have
really different palette from it
- it ignores stylesheets (or modifications to the QLineEdit's palette)

Therefore, the color for text format is picked up from focusWidget,
if it is available (in most cases it should be). If the focusWidget
is not available, then QLineEdit default palette is used.

Task-number: QTBUG-9480
Reviewed-by: Janne Koskinen