QMetaObject::invokeMethod using Qt::BlockingQueuedConnection can handle the return value.
When using Qt::BlockingQueuedConnection, we do not need to copy the
arguments, and we can handle the return type.

The argv we pass to the event is the param vector alocated on the
stack. Since we don't need to destroy the argument, we can pass 0
for the types.
The private QMetaCallEvent destructor is modified not to destroy
the arguments if types_ is 0

Task-number: QTBUG-10440
Reviewed-by: Brad