Fixed bug in QIODevice::read after first reading 0 bytes.
Change 02532ec80375c686503c4250c6ad6bb211515ec8 removed the early-exit
for 0 byte reads, causing us to hit code that assumed the buffer was
empty since nothing was read. It would thus read more into the end of
the buffer, causing the buffer to grow bigger than QIODEVICE_BUFFERSIZE.
Next, if the actual number of bytes we wanted to read was bigger than
the original buffer size we'd read the same data twice.

Reviewed-by: João Abecasis
Reviewed-by: Thiago Macieira