Fixed scrolling bugs in widget graphics effect backend.
The cache wasn't invalidated for widgets with a graphics effect inside a
scrolling QScrollArea, so for now we disable caching for widget source
pixmaps. Also, we can't clip the source pixmap to device coordinates
since there's no knowledge of which areas of the source pixmap the
effect uses to compute the visible pixels. See change fd30cc9fabe6fc023
for the graphics item effect backend fix.

Reviewed-by: Bjørn Erik Nilsen