Fix off-by-one in text layouts and widget size hints on Mac
The mac font engine was made to return fractional values. This has
exposed a few bugs where the rounded height of the font was used for
text layouts and widget size hints. Since a fractional part of the
font's height also creates pixels, rounding the height down in these
cases is always wrong.

The biggest culprit was QScriptLine::height. This is used to lay out
text lines and returning a fractional height created problems all
over the place. Rather than ceil the number in all places where it
is used, we simply return a ceiled value for the line's height.
In 4.6, this value would be qRound(ascent)+qRound(descent)+1,
which would in some cases cause the previous and current line to
overlap by one pixel. By ceiling the sum of the two, we guarantee
that the text line contains the complete bounds of the font.

A note about the removal of the code that gets the font metrics for
QSmallFont: The condition would never be true since this is inside
a branch which is only evaluated if fontIsSet is true.

Task-number: QTBUG-9971
Reviewed-by: Gunnar