Implement UDP multicast on Windows
The code is very similar to the UNIX version, but not identical. It
appears the Microsoft compiler does not like the "interface" variable
name, so they have been renamed.

The autotest had a bug in it that I failed to notice before. The sender
was not sending to the multicast group, it was sending to the local
address of the receiver (which fails on Windows).