Really fix tst_QDockWidget::taskQTBUG_9758_undockedGeometry on Linux
If there is no window manager the current coordinate are taken in account
even if WA_Moved is false.
When a dockwidget is hidden, it is sent far in the negative coordinate.

The test was failing because the WM is assync and may take some time to
update the geometry.