Don't crash when applications set Qt::WA_TranslucentBackground.
After replacing some code with a function call to s60UpdateIsOpaque()
we introduced a crash for widgets that set
Qt::WA_TranslucentBackground. The reason for the crash was that the
Qt::WA_WState_Created attribute was being set before calling this
function, but *not* before setWinId() was being called. This meant that
s60UpdateIsOpaque() assumed that the window was created (and the handle
added to the widget map), but this was not actually the case.

The fix here is to move the call to s60UpdateIsOpaque() after the call
to setWinId() such that those assumptions are true.

Reviewed-by: Jani Hautakangas