New Pad Nagivator example implementation.
The old implementation was demo-quality and didn't really belong in
the examples section. This new code cleans up the implementation, and
also makes use of brand new features such as the state machine and
animated transitions system.

The new code is more suitable for documentation, and full docs are
coming up in the next commit.