Fix crash in ParentAnimation.
copyOriginals plays with the order of the revertList, which messes up
the assumptions of ParentAnimation. The full fix will require some
rearchitecting of how the states and transitions handle "related"
actions, but for now this fixes the crash.

Reverting to the base state has also been fixed.

Task-number: QTBUG-10671, QTBUG-10676