Mac: restoreGeometry()
The problem here is the fact that when zooming (maximizing) a window,
we should update the starting position since the window is now starting
at 0,0. However when restoring the window should be placed in the
original position. The bug here occurs because we kept the original
position and the new size. So the window was drawn were it was using
the new size. This patch introduces a boolean flag that keeps track of
the maximized state. We are interested in knowing if the window was
maximized but not restored, so when saving the geometry we have a way
to know that we need to "move" the window to 0,0.

Task-number: QTBUG-10504
Reviewed-by: Richard Moe Gustavsen