QS60Style: In S60 3.x and 5.0 Qt itemviews behaviour is not nativelike
Currently style defines QStyle::SH_ItemView_ActivateItemOnSingleClick,
which leads to itemview item activation only occur on double-click
(or tap). The item selection work native-like on Sym^3.

As a correction, style needs to check the passed styleoption and
check whether or not the state "Selected" is on. In these cases,
the itemactivation can proceed.

Task-number: QTBUG-10697
Reviewed-by: Miikka Heikkinen