define qtPrepareTool() function and use it throughout
the function provides a cross-platform way to determine the exact
pathname of our build tools (moc, etc.).

use it in our .prf files, so we don't have to rely on qmake's
unreliable path separator normalization magic in extra compiler
commands, which broke on mingw+sh in silent mode.

remove the bootstrap tool path setting from configure, as it is
redundant now.

Reviewed-by: joerg
Task-number: QTBUG-10633