Support linked fonts (.ltt) from standard font locations.
The internal font database was only populated with .ttf and .ccc files.
This patch adds .ltt as file type.

Without complete font type coverage in the fontstore, we may get a
mismatch in the internal association of font table and fontfamily.

Most probably, this will also fix the crash on SSE Satio's. SSE
seems to use .ltt files already on S60 5.0. An improvement needs to be
verified by an owner of such a device, however.

Task-number: QTBUG-8905
Reviewed-by: Aleksandar Sasha Babic