Fix for autotest failure in qwidget::saveRestoreGeometry()
The problem was that the fix for bug QTBUG-10519 introduced a new state
for saving the geometry. However there is an exception and that
exception comes into play when saving a maximized window. This fixes
that by making sure we only enter into that state when the window is not
currently maximized.

Reviewed-by: tbastian