Limit the lower glyph texture cache size to 16x16 in GL.
The SGX chip has a limitation regarding FBOs in that it doesn't support
FBO sizes that are POT and less than 16 pixels in width/height.
Since the FBO we use in the GL glyph cache always mirrors the size
of the texture itself, we limit that instead to avoid creating
invalid FBOs.

Task-number: QTBUG-10645
Reviewed-by: Gunnar