qmake: added possibility to specify the type of an install target
Before this change:
target.CONFIG+=no_check_exist implies the file is a file, no way to
make an install rule for a non-existing directory.

Now, its possible to specify the type:
    target.CONFIG+=no_check_exist directory
will install a directory.
    target.CONFIG+=no_check_exist executable
will install an executable.
    target.CONFIG+=no_check_exist data
will install a normal file.

The default case, if no type is given, like in
    CONFIG+=no_check_exist
will call QFileInfo::isExecutable() to determine, if its a data file
or executable. This is the old behaviour.

Task-number: QTBUG-10624
Reviewed-by: ossi