Fixed tst_compilerwarnings test failure due to icecc node failures.
When an icecc node has some system error, icecc on the client side
outputs a warning. Since it's able to recover and the warning has
nothing to do with the code, we should ignore it.

Other distributed compile tools will have similar issues, but no attempt
has been made to cover them.