Small hack to allow forcing off font smoothing on Mac OS X
In declarative all text is cached in a transparent buffer. Using subpixel
antialiasing for this gives subpar results, and for better results,
we want to turn off the font smoothing in the OS for this case. Since
we cannot efficiently detect non-opaque destination pixels, we have to
have a way where a user can override the default. We'll introduce a
style strategy for this later, but for now we only want to enable the
use of CGContextSetShouldSmoothFonts() to explicitly turn smoothing off
in the declarative context. We still override if font smoothing has been
explicitly turned off by us for the context by the TextAntialiasing
flag.

Task-number: QT-2776
Reviewed-by: MortenS