Improve virtual mouse on E72 optical joystick
Uses scancodes instead of keycodes, to get access to the 8-way cursor
keys introduced in S60 3.2. (EStdKeyDevice10-13)

Doubled the rate of acceleration and top speed of the cursor (this
applies to all phones, but feedback was that the cursor was too slow)

Multiple key presses within 500ms of each other are treated the same as
key repeat events. This is because the E72 optical joystick sends pulsed
key presses.

Task-number: QTBUG-5711
Reviewed-by: Sami Merila