QApplication::closeAllWindows() should ignore windows being closed
It is very common to display a dialog in response to a close event.
Closing the window again will result in QWidget::close() returning true.
This confuses QApplication::closeAllWindows(), since the window is still
visible even though it was closed (or is closing). Solve this by
ignoring windows that have the is_closing flag set in their widget data.

Task-number: QTBUG-7635
Reviewed-by: Denis Dzyubenko