QPushButton's click area exceeds the button area by far on Mac OS X
The problem is the "rounding" of buttons in OSX. To achieve this we
add some padding to the buttons, therefore creating an inner rect for
the widget. The common hitButton(...) method found in QAbstractButton
just checks the normal rect of the widget. What this patch does is to
reimplement hitButton(...) in QPushButton, but only for the Mac case.
In this reimplemented method I calculate the inner rect and check if
the hit point is inside that rect or not.

Task-number: QTBUG-10401
Reviewed-by: Richard Moe Gustavsen