Fix crash when using fonts in non-gui QApplication
XLFD requires a DISPLAY connection to work. We would get crashes in
a QApplication with GUI disabled when: 1. Not using FontConfig, or
2. Falling back to XLFD for bitmap fonts that require scaling.

The patch disables paths to loadXlfd() when GUI is disabled. This means
that in XLFD, we will always get a box font, which is the same behavior
as when using fonts outside the main thread. There doesn't seem to be
any way around this.

With FontConfig, we will use the font it returns, even if it's a
slightly wrong size. Main consequence will be for using bitmap
fonts for printing on a highres printer in a non-gui application.
Again, there does not seem to be any way around this.

NOTE: I've also added a catch to avoid going into loadXlfd() in the
fallback if we're not on the main thread, since this was missing.

Task-number: QTBUG-10448
Reviewed-by: Trond