Improve look and feel of itemviews on mac
These fixes are neccessary and requested by Qt Creator and
should go into 4.7.0. The fixes involved is:

 - positioning of disclosure arrow in treeviews
 - spacing between icon and text label
 - added 2 pixel spacing between item view items

Note that native cocoa views tend to use 3 pixel spacing.
We considered this and decided on leaving it at 2 which is what
you see in finder list views, XCode among other places.
It also makes the change a bit less radical.

Reviewed-by:cduclos
Task-number:QTBUG-10190