Integrate some QML examples and demos into qtdemo
Includes minor changes and additions to the existing doc and examples,
so that they follow Qt conventions better.

Note that while blurring the background was part of the plan for the
embedded QML viewer I could not get it to perform well enough. In the
future, when blur is fast enough (or someone else can get it to
perform better than I) -use-blur should become the default, and
-no-blur the option.

Task-number: QTBUG-10582