Qt app draws background incorrectly when animated wallpaper is used
In Symbian^3 OS supports animated wallpapers. Unfortunately animation
is drawn to separate surface underneath the application surface. To
avoid fiddling with system surfaces, Qt apps indicate to the
OS that application does not support animated wallpaper and thus,
system draws a "regular" wallpaper for the application.

This feature is supported only in Sym^3.

Task-number: QT-3148
Reviewed-by: Shane Kearns