Update the state of the keyboard modifiers on drop events.
In the case of drag-n-drop between processes we do not get key events
for the pressed modifiers, so we need to update the state of the modifiers
(using XQueryPointer) to get a correct value of QApplication::keyboardModifiers()
and QDropEvent::keyboardModifiers() at drop time.

This is necessary for fixing KDE bug 178679.

Merge-request: 590
Reviewed-by: Bradley T. Hughes <bradley.hughes@nokia.com>
Reviewed-by: Andreas Aardal Hanssen <andreas.aardal.hanssen@nokia.com>