Added missing PlatSec capabilities to spectrum demo DLL
On Symbian, spectrum.exe has UserEnvironment capability in order to
allow it to use QAudioInput.  This means that the capabilities of
any DLL which it loads must include UserEnvironment.  Prior to this
patch, fftreal.dll was built without any capabilities, causing
startup of the application to fail.

Although granting only UserEnvironment to fftreal.dll would fix the
bug, the usual pattern on Symbian OS is to grant 'all -tcb'
capabilities to DLLs.  The capabilities with which the DLL actually
runs are inherited from its parent process.

Task-number: QTBUG-10964
Reviewed-by: Liang Qi