Redraw issues when removing a fully transparent QGraphicsItem from the scene.
This only happened with fully transparent ancestors (item's
effectiveOpacity() == 0.0). Problem was that we didn't take into
account the ancestors' opacity when removing an item from the scene.
More specifically: The calculated effective opacity for the item was
zero and we ignored update requests. We have to ignore the opacity if
any of the ancestors' ignoreOpacity bit is 1, which means the opacity
is set to 0 and the update request has not yet been processed.

Auto test included.

Task-number: QTBUG-10778