Improve precision of testlib benchmarking.
20 ms is a bit too to get a precise measurement due to limited timer
resolution. We increase the limit to 50 ms and introduce a warm-up iteration
to prevent any state or cache-related issues which can be hard to account
for in the tests.

On the N900 this seems to decrease the variance of test results by 30 %,
and fixes some cases where the results were way off due to hidden state.

Task-number: QT-3390
Reviewed-by: Morten Sørvig