Fix multithreaded use of bearer management on Symbian.
Move more initialization code into initialize() function.
The constructor for the SymbianEngine class can be called from any
thread. Move all initialization into initialize() function which is
guarinteed to be run in the main thread. This is required as some
Symbian resources are not shared between thread in the same process.

The implementation of the defaultConfiguration() function could only be
called from the main thread. Cache a copy of the default network
configuration when the configurations are updated and just return this
when the defaultConfiguration() function is called.

Only lock the engine mutex from external entry points.

Remove need for locking in waitRandomTime() function. Prevents dead
lock due to nested event loop.

Task-number: QTBUG-10296