Cocoa: remove 'double' content view from window
Up til now, we always used to create a new NSView for a top
level widget, and make it a child of the windows content view.
This is/should be unnecessary, as NSWindow has a method for
setting the content view directly, thereby allowing for reducing
the stacking level with one view. Not the biggest gain if the
year, but it seems effortless to implement. And I do it as
a part of Alien preparation.

Rev-By: msorvig