Fix unreasonably large width of QTextLayout::boundingRect()
When no lineWidth is set on the QTextLine, the QScriptLine::width will
be the maximum value that can be represented by QFixed, thus always
overriding the width of the bounding rect. The lineWidth of the
QTextLine should never be used as the width of the bounding rect, since
it's not a calculated value. The real bounding rect is found by querying
the natural text width.

Task-number: QTBUG-11104
Reviewed-by: Rhys Weatherley