Make symbian install to epocroot by default.
On symbian cross-compiling is the only way so the libraries and headers
are all stored in symbian-specific ways which is called EPOCROOT.
This change makes a configure without -prefix being passed in populate
the EPOCROOT on 'make install'.
When the user passes a prefix it will just do what a normal unix build
does on make install.