Revert behavior of QTextLayout::boundingRect() when line width is set
In change 817469519a784b1cc84f89cb3cb84f7560874f8e, there was a
behavioral change which can cause regressions, as the bounding rect of
the QTextLayout would previously return the set line width when this
was greater than the calculated natural text width. We revert to this
behavior to avoid regressions and add an autotest for it. When the
line width is not set (and si.width is equal to QFIXED_MAX), then we
will still return the natural text width.

Reviewed-by: Lars