Add some convenience functions for QStringRef
QStringRef was hard to use for a lot of real-world tasks since it lacked
methods similar to the ones in QString for
indexOf/lastIndexOf/contains/startsWith/endsWith

This patch adds that functionality.

I also copied the relevant test cases from QString for the new APIs and
introduced a QStringRef test.

Merge-request: 625
Reviewed-by: Olivier Goffart <olivier.goffart@nokia.com>
Reviewed-by: ossi