fix setRawData()
setRawData() itself (consistently with fromRawData()) sets alloc to the
passed raw size, so the check at the start would always find that we
cannot re-use the qstring object, thus completely defeating the purpose
of the function.
so also check that the string data pointer actually points into the
internal array before giving up.

Reviewed-by: olivier