Avoid unnecessary detach / deep copy in QGraphicsItem::scroll().
QGraphicsItem::scroll() scrolls the cache pixmap when cacheMode is
enabled (for ItemCoordinateCache only). Because the pixmap exists both
in the cache and in a temp QPixmap copy, the ref count is 2, so the
scroll operation has to do a deep copy. To avoid this, we remove the
pixmap from the cache before and reinsert it again after calling
QPixmap::scroll().

Reviewed-by: Alexis Menard <alexis.menard@nokia.com>