Fix bug in QGraphicsItem::scroll() when called with no QRectF argument.
With a null rect argument, QGraphicsItem::scroll() is supposed to scroll
the whole item. When using ItemCoordinateCache, currently the only
supported mode for scroll optimization, we simple scrolled the whole
cache pixmap. Problem: The cache pixmap has a border of 2 pixels. So we
scroll the contents _and_ the border. This leaves white/transparent
horizontal and vertical line artifacts when scrolling.

This change unifies the two cases - partial and full scrolling - into
one (shorter) approach that works without scrolling the margin as well.

Reviewed-by: Alexis Menard <alexis.menard@nokia.com>